[chore] move opaque confmap test
This removes the dependency on confmap for tests in configopaque
