+++ config/configtls/go.mod
@@ -17,5 +17,3 @@ require (
 )
 
 replace go.opentelemetry.io/collector/config/configopaque => ../configopaque
-
-replace go.opentelemetry.io/collector/confmap => ../../confmap
